scratch/test_imports.py
try:
    from google.adk.agents import Agent
    from google.adk.runners import Runner
    from google.adk.sessions import InMemorySessionService
    from google.adk.tools.bigquery import BigQueryToolset, BigQueryCredentialsConfig
    print("Success: All ADK modules imported!")
except ImportError as e:
    print(f"ImportError: {e}")
